% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tty.R
\name{extended_product_ttys}
\alias{extended_product_ttys}
\title{Extended RxNorm product term types}
\usage{
extended_product_ttys()
}
\value{
A character vector of RxNorm term type abbreviations.
}
\description{
Returns an extended set of RxNorm term types that includes the default
product term types plus dose-form and dose-form-group concepts.
}
\details{
This can be useful when a broader set of product-related RxNorm concepts
is desired.
}
\examples{
extended_product_ttys()

}
