% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tty.R
\name{extended_ttys}
\alias{extended_ttys}
\title{Extended RxNorm term types}
\usage{
extended_ttys()
}
\value{
A character vector of RxNorm term type abbreviations.
}
\description{
Returns an extended set of RxNorm term types that includes the default and extended
product term types plus drug component, brand name, and ingredient concepts.
This list includes essentially all term types that capture a specific ingredient,
thus it excludes dose form, dose form group, prescribable name, synonyms, and others
that are not associated with a specific ingredient.
}
\details{
This can be useful when the broadest set of RxNorm concepts that still capture an ingredient
is desired.
}
\examples{
extended_ttys()

}
