% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{find_classes}
\alias{find_classes}
\title{Find RxClass drug classes by class name}
\usage{
find_classes(query, class_types = NULL)
}
\arguments{
\item{query}{Character string to search for.}

\item{class_types}{Optional class type filter, such as \code{"EPC"}, \code{"ATC1"},
\code{"ATC2"}, \code{"ATC3"}, \code{"ATC4"}, \code{"CHEM"}, \code{"DISEASE"}, or \code{"VA"}.}
}
\value{
A tibble of matching class concepts.
}
\description{
Find RxClass drug classes by class name
}
