% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_class_members}
\alias{get_class_members}
\title{Get RxNorm members of an RxClass class}
\usage{
get_class_members(
  class_id,
  rela_source,
  rela = NULL,
  include_indirect = TRUE,
  ttys = NULL
)
}
\arguments{
\item{class_id}{RxClass class identifier.}

\item{rela_source}{RxClass relationship source, such as \code{"ATC"},
\code{"ATCPROD"}, \code{"DAILYMED"}, \code{"FDASPL"}, \code{"MEDRT"}, \code{"SNOMEDCT"}, or \code{"VA"}.}

\item{rela}{Optional relationship filter.}

\item{include_indirect}{Logical; if \code{TRUE}, include direct and indirect
members. If \code{FALSE}, return direct members only.}

\item{ttys}{Optional RxNorm term type filter, such as \code{"IN"}, \code{"PIN"},
\code{"SCD"}, or \code{"SBD"}.}
}
\value{
A tibble of RxNorm drug members.
}
\description{
\code{get_class_members()} queries the RxClass API to identify members of a
specific drug class. Typically, this would be used to identify all
ingredient RxCUIs in a class, which can then be combined with other search
functions to extract related product RxCUIs or NDCs.
}
