% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{get_properties}
\alias{get_properties}
\title{Get core RxNorm properties for one or more RxCUIs}
\usage{
get_properties(rxcui, show_progress = interactive())
}
\arguments{
\item{rxcui}{Character vector of RxCUIs}

\item{show_progress}{Logical. Show a progress bar in interactive sessions.}
}
\value{
tibble with rxcui, name, synonym, tty, language, suppress, umlscui.
}
\description{
Get core RxNorm properties for one or more RxCUIs
}
\examples{
if (identical(Sys.getenv("RXREF_ONLINE"), "1")) {
get_properties(c("860975","1049630"))
}
}
