% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tty.R
\name{product_ttys}
\alias{product_ttys}
\title{RxNorm product term type sets}
\usage{
product_ttys(set = c("default", "extended_product", "extended"))
}
\arguments{
\item{set}{One of \code{"default"}, \code{"extended_product"}, or \code{"extended"}.}
}
\value{
A character vector of RxNorm term type abbreviations.
}
\description{
Returns common RxNorm term type sets used by rxref.
}
\examples{
product_ttys()
product_ttys("extended")

}
