% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rxclass_rela_sources}
\alias{rxclass_rela_sources}
\title{Get RxClass relationship sources}
\usage{
rxclass_rela_sources()
}
\value{
A tibble of RxClass relationship sources.
}
\description{
Get RxClass relationship sources
}
