% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rxclass_relas}
\alias{rxclass_relas}
\title{Get RxClass relationship types}
\usage{
rxclass_relas(rela_source)
}
\arguments{
\item{rela_source}{RxClass relationship source, such as \code{"ATC"},
\code{"DAILYMED"}, \code{"FDASPL"}, \code{"MEDRT"}, \code{"SNOMEDCT"}, or \code{"VA"}.}
}
\value{
A tibble of relationship names for the selected source.
}
\description{
\code{rxclass_relas()} is a helper function that queries the RxClass API for
generally valid relationships within the data. Note that just because a
relationship is generally valid within a rela_source, not every drug or
RxCUI will have a corresponding relationship asserted.
}
