% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_backend}
\alias{set_backend}
\title{Select backend (API-only for now)}
\usage{
set_backend(backend = c("api"))
}
\arguments{
\item{backend}{Character scalar; only "api" is currently supported.}
}
\description{
Select backend (API-only for now)
}
