% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tty.R
\name{tty_catalogue}
\alias{tty_catalogue}
\title{Catalogue of RxNorm TTY (Term Types)}
\usage{
tty_catalogue()
}
\value{
A tibble with metadata for key TTYs.
}
\description{
Returns a tibble describing common RxNorm TTYs you may want to use when
expanding ingredients to products. Includes whether each TTY typically maps
cleanly to NDCs and whether it's included in the package's default or
extended TTY sets.
}
\details{
Columns:
\itemize{
\item \code{tty}: RxNorm term type code.
\item \code{label}: Short, human-friendly name.
\item \code{description}: What the TTY represents in RxNorm.
\item \code{maps_to_ndc}: Logical; whether CUIs of this TTY usually map to NDCs via
\verb{/rxcui/\{rxcui\}/ndcs}.
\item \code{typical_role}: How it’s commonly used (product, component, group, name).
\item \code{include_default}: Logical; included in \code{.rxref_default_ttys}.
\item \code{include_extended}: Logical; included in \code{.rxref_extended_ttys}.
}
}
\examples{
tty_catalogue()

# TTYs that map to NDCs
subset(tty_catalogue(), maps_to_ndc)$tty

# See what your defaults and extended sets contain
subset(tty_catalogue(), include_default)$tty
subset(tty_catalogue(), include_extended)$tty

# Pick a custom set: products + components
with(tty_catalogue(),
     tty[tty \%in\% c("SCD","SBD","GPCK","BPCK","SCDC","SBDC")])
}
