% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{get_clinical_attributes}
\alias{get_clinical_attributes}
\title{Clinical attributes from the concept (SCD/SBD) or related SCD/SBD}
\usage{
get_clinical_attributes(rxcui)
}
\arguments{
\item{rxcui}{Character vector of RxCUIs}
}
\value{
A tibble with 6 columns: \code{rxcui}, \code{related_rxcui}, \code{name}, \code{tty}, \code{strength}, \code{dose_form}
}
\description{
If \code{rxcui} is already a clinical drug (SCD/SBD), parse attributes directly
from its own name. Otherwise, query \code{/rxcui/{id}/related?tty=SCD,SBD} and parse.
}
\details{
Note: This works, but is still in active development.
}
\examples{
\donttest{
get_clinical_attributes(c("861007","860975")) |>
  dplyr::select(rxcui, related_rxcui, name, strength, dose_form, tty) |>
  head()
}
}
