% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_contraindications}
\alias{get_contraindications}
\title{Get contraindication assertions from MED-RT}
\usage{
get_contraindications(x, by = c("rxcui", "name"), keep_input = TRUE)
}
\arguments{
\item{x}{Character vector of RxCUIs or drug names.}

\item{by}{One of \code{"rxcui"} or \code{"name"}.}

\item{keep_input}{Logical; if \code{TRUE}, includes the original input value.}
}
\value{
A tibble of MED-RT contraindication assertions.
}
\description{
\code{get_contraindications()} is a convenience wrapper for \code{get_medrt()} that
returns only contraindication MED-RT assertions for RxNorm drugs.
}
