% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_may_treat}
\alias{get_may_treat}
\title{Get indication/treatment assertions from MED-RT}
\usage{
get_may_treat(x, by = c("rxcui", "name"), keep_input = TRUE)
}
\arguments{
\item{x}{Character vector of RxCUIs or drug names.}

\item{by}{One of \code{"rxcui"} or \code{"name"}.}

\item{keep_input}{Logical; if \code{TRUE}, includes the original input value.}
}
\value{
A tibble of MED-RT may-treat assertions.
}
\description{
\code{get_may_treat()} is a convenience wrapper for \code{get_medrt()} that returns
indications (disease assertions) from MED-RT for RxNorm drugs, when
available. Note that absence of a row (or an empty table altogether)
should not necessarily be interpreted as absence of an indication.
}
