% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_medrt_pe}
\alias{get_medrt_pe}
\title{Get MED-RT physiologic-effect assertions}
\usage{
get_medrt_pe(x, by = c("rxcui", "name"), keep_input = TRUE)
}
\arguments{
\item{x}{Character vector of RxCUIs or drug names.}

\item{by}{One of \code{"rxcui"} or \code{"name"}.}

\item{keep_input}{Logical; if \code{TRUE}, includes the original input value.}
}
\value{
A tibble of MED-RT physiologic-effect assertions.
}
\description{
\code{get_medrt_pe()} is a convenience wrapper for \code{get_medrt()} that returns
only physiologic effect MED-RT assertions for RxNorm drugs.
}
