% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_to}
\alias{map_to}
\alias{map_ndc_to_rxcui}
\alias{map_rxcui_to_ndc}
\title{Map between NDC and RxCUI in either direction}
\usage{
map_to(x, to = c("rxcui", "ndc"), status = NULL, show_progress = interactive())

map_ndc_to_rxcui(x, show_progress = interactive())

map_rxcui_to_ndc(x, status = NULL, show_progress = interactive())
}
\arguments{
\item{x}{Character vector of NDCs (10/11-digit or hyphenated) or RxCUIs.}

\item{to}{One of c("rxcui","ndc") indicating desired output id.}

\item{status}{For \code{to = "ndc"}, filter by NDC status, e.g., "ACTIVE".
Use NULL for all.}

\item{show_progress}{Logical. Show a progress bar in interactive sessions.
Progress is shown only when at least 5 inputs are supplied.}
}
\value{
tibble with columns depending on direction.
}
\description{
Map between NDC and RxCUI in either direction

Convenience wrappers
}
