% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{ndc_status}
\alias{ndc_status}
\title{Get RxNav status for NDCs}
\usage{
ndc_status(ndc)
}
\arguments{
\item{ndc}{character vector; hyphenated or digits}
}
\value{
A tibble with three columns: ndc_input, ndc11, ndc_hyph, status
}
\description{
\code{ndc_status()} requests the status of an NDC entry for one or more user-supplied
NDC numbers. The returned status will be one "ACTIVE", "OBSOLETE", or "UNSPECIFIED".
On occasion, a NULL value may be pulled (resulting in an \code{NA} in the returned tibble),
when RxNorm has no status for a given NDC, or if an invalid NDC is supplied.
}
