% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{ndc_status}
\alias{ndc_status}
\title{Get RxNav status for NDCs}
\usage{
ndc_status(ndc)
}
\arguments{
\item{ndc}{character vector; hyphenated or digits}
}
\value{
tibble ndc_input, ndc11, ndc_hyph, status
}
\description{
Get RxNav status for NDCs
}
