% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{rxref-package}
\alias{rxref}
\alias{rxref-package}
\title{rxref: Tidy RxNorm utilities (API-first)}
\description{
Core exported verbs: \code{\link[=resolve]{resolve()}}, \code{\link[=get_properties]{get_properties()}}, \code{\link[=map_to]{map_to()}}.
Configure behavior with \code{\link[=rxref_conf]{rxref_conf()}} and \code{\link[=set_backend]{set_backend()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ssmithm/rxref}
  \item Report bugs at \url{https://github.com/ssmithm/rxref/issues}
}

}
\author{
\strong{Maintainer}: Steven Smith \email{ssmith@cop.ufl.edu}

}
\keyword{internal}
