% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{rxref_conf}
\alias{rxref_conf}
\title{Configure rxref}
\usage{
rxref_conf(
  base_url = getOption("rxref.base_url"),
  rate_delay = getOption("rxref.rate_delay"),
  cache = getOption("rxref.cache")
)
}
\arguments{
\item{base_url}{Override the RxNav base URL (e.g., a local mirror)}

\item{rate_delay}{Seconds to wait between HTTP calls}

\item{cache}{A cachem cache object used by memoised calls}
}
\value{
A named list of current settings
}
\description{
Configure rxref
}
